requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
bcrypt==3.2.0
bingscraper==3.0
boto==2.49.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
distro==1.5.0
Django==3.1
django-dynamodb-sessions==0.6
djangorestframework==3.11.1
docker==4.3.0
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.5.2
paramiko==2.7.1
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dotenv==0.14.0
pytz==2020.1
pywin32==227
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
Scrapy==2.3.0
scrapy-djangoitem==1.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
texttable==1.6.2
Twisted @ file:///C:/Users/luxmicom/Desktop/ARSH%20DJANGO%20PROJECT/test_project/local-env/Lib/site-packages/Twisted-20.3.0-cp38-cp38-win_amd64.whl
urllib3==1.25.10
w3lib==1.22.0
websocket-client==0.57.0
zope.interface==5.1.0
